Extracted the role API into it's own file
`installerRole` is now `model.Role`.
